+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/api/BinaryModuleApiService.java
@@ -24,18 +24,18 @@ public interface BinaryModuleApiService {
 
     ResponseEntity<String> getPackageBinaryModules(String package_name,
                                                    String package_version,
-                                                   short offset,
-                                                   short limit);
+                                                   int offset,
+                                                   int limit);
 
     ResponseEntity<String> getBinaryModuleMetadata(String package_name,
                                                    String package_version,
                                                    String binary_module,
-                                                   short offset,
-                                                   short limit);
+                                                   int offset,
+                                                   int limit);
 
     ResponseEntity<String> getBinaryModuleFiles(String package_name,
                                                 String package_version,
                                                 String binary_module,
-                                                short offset,
-                                                short limit);
+                                                int offset,
+                                                int limit);
 }
