+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/api/EdgeApiService.java
@@ -24,6 +24,6 @@ public interface EdgeApiService {
 
     ResponseEntity<String> getPackageEdges(String package_name,
                                            String package_version,
-                                           short offset,
-                                           short limit);
+                                           int offset,
+                                           int limit);
 }
