+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/api/ModuleApi.java
@@ -34,8 +34,8 @@ public class ModuleApi {
     @GetMapping(value = "/{pkg}/{pkg_ver}/modules", produces = MediaType.APPLICATION_JSON_VALUE)
     ResponseEntity<String> getPackageModules(@PathVariable("pkg") String package_name,
                                              @PathVariable("pkg_ver") String package_version,
-                                             @RequestParam(required = false, defaultValue = "0") short offset,
-                                             @RequestParam(required = false, defaultValue = RestApplication.DEFAULT_PAGE_SIZE) short limit) {
+                                             @RequestParam(required = false, defaultValue = "0") int offset,
+                                             @RequestParam(required = false, defaultValue = RestApplication.DEFAULT_PAGE_SIZE) int limit) {
         return service.getPackageModules(package_name, package_version, offset, limit);
     }
 
@@ -43,8 +43,8 @@ public class ModuleApi {
     ResponseEntity<String> getModuleMetadata(@PathVariable("pkg") String package_name,
                                              @PathVariable("pkg_ver") String package_version,
                                              @RequestBody String module_namespace,
-                                             @RequestParam(required = false, defaultValue = "0") short offset,
-                                             @RequestParam(required = false, defaultValue = RestApplication.DEFAULT_PAGE_SIZE) short limit) {
+                                             @RequestParam(required = false, defaultValue = "0") int offset,
+                                             @RequestParam(required = false, defaultValue = RestApplication.DEFAULT_PAGE_SIZE) int limit) {
         return service.getModuleMetadata(package_name, package_version, module_namespace, offset, limit);
     }
 
@@ -52,8 +52,8 @@ public class ModuleApi {
     ResponseEntity<String> getModuleFiles(@PathVariable("pkg") String package_name,
                                           @PathVariable("pkg_ver") String package_version,
                                           @RequestBody String module_namespace,
-                                          @RequestParam(required = false, defaultValue = "0") short offset,
-                                          @RequestParam(required = false, defaultValue = RestApplication.DEFAULT_PAGE_SIZE) short limit) {
+                                          @RequestParam(required = false, defaultValue = "0") int offset,
+                                          @RequestParam(required = false, defaultValue = RestApplication.DEFAULT_PAGE_SIZE) int limit) {
         return service.getModuleFiles(package_name, package_version, module_namespace, offset, limit);
     }
 
@@ -61,8 +61,8 @@ public class ModuleApi {
     ResponseEntity<String> getModuleCallables(@PathVariable("pkg") String package_name,
                                               @PathVariable("pkg_ver") String package_version,
                                               @RequestBody String module_namespace,
-                                              @RequestParam(required = false, defaultValue = "0") short offset,
-                                              @RequestParam(required = false, defaultValue = RestApplication.DEFAULT_PAGE_SIZE) short limit) {
+                                              @RequestParam(required = false, defaultValue = "0") int offset,
+                                              @RequestParam(required = false, defaultValue = RestApplication.DEFAULT_PAGE_SIZE) int limit) {
         return service.getModuleCallables(package_name, package_version, module_namespace, offset, limit);
     }
 }
