+++ analyzer/restapi-plugin/src/main/java/eu/fasten/analyzer/restapiplugin/mvn/api/impl/BinaryModuleApiServiceImpl.java
@@ -30,8 +30,8 @@ public class BinaryModuleApiServiceImpl implements BinaryModuleApiService {
     @Override
     public ResponseEntity<String> getPackageBinaryModules(String package_name,
                                                           String package_version,
-                                                          short offset,
-                                                          short limit) {
+                                                          int offset,
+                                                          int limit) {
         String result = KnowledgeBaseConnector.kbDao.getPackageBinaryModules(
                 package_name, package_version, offset, limit);
         result = result.replace("\\/", "/");
@@ -42,8 +42,8 @@ public class BinaryModuleApiServiceImpl implements BinaryModuleApiService {
     public ResponseEntity<String> getBinaryModuleMetadata(String package_name,
                                                           String package_version,
                                                           String binary_module,
-                                                          short offset,
-                                                          short limit) {
+                                                          int offset,
+                                                          int limit) {
         String result = KnowledgeBaseConnector.kbDao.getBinaryModuleMetadata(
                 package_name, package_version, binary_module, offset, limit);
         result = result.replace("\\/", "/");
@@ -54,8 +54,8 @@ public class BinaryModuleApiServiceImpl implements BinaryModuleApiService {
     public ResponseEntity<String> getBinaryModuleFiles(String package_name,
                                                        String package_version,
                                                        String binary_module,
-                                                       short offset,
-                                                       short limit) {
+                                                       int offset,
+                                                       int limit) {
         String result = KnowledgeBaseConnector.kbDao.getBinaryModuleFiles(
                 package_name, package_version, binary_module, offset, limit);
         result = result.replace("\\/", "/");
